AHM-7 Padronização dos textos na aplicação, remoção de códigos inutilizados e salvar filtros utilizados.

+++ Stefanini.ViaReport.Core/Data/Dto/DashboardInfoItemDto.cs
@@ -6,7 +6,7 @@ namespace Stefanini.ViaReport.Core.Data.Dto
     public class DashboardInfoItemDto
     {
         public DateTime Date { get; set; }
-        public long Value { get; set; }
+        public decimal Value { get; set; }
         public IList<IssueInfoDto> Issues { get; set; }
 
     }
